scripts/requirements.txt
gspread==6.1.4
google-auth==2.38.0
python-dotenv==1.1.0
